Move timeslot functionality into its own utils file
